clen code

+++ src/components/Myideas.tsx
@@ -21,61 +21,56 @@ import {
 import React from "react";
 import Footer from "./Footer";
 import Nav from "./Nav";
-import { useNavigate ,Link} from "react-router-dom";
+import { useNavigate, Link } from "react-router-dom";
 
 function Myideas() {
   const [data, setData] = React.useState<any[]>([]);
-  const navigate = useNavigate()
+  const navigate = useNavigate();
 
- 
-    // fetch data
-    const getallproject = async () => {
-      const data = await (
-        await fetch(
-          "http://localhost:3008/idea/",{
-            method: "GET",
-            headers: {
-              "Content-Type": "application/json",
-              "authorization": localStorage.getItem('token') as string,
-            },
-          }
-        )
-      ).json();
+  // fetch data
+  const getallproject = async () => {
+    const data = await (
+      await fetch("http://localhost:3008/idea/", {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+          authorization: localStorage.getItem("token") as string,
+        },
+      })
+    ).json();
 
-      // set state when the data received
-      setData(data&&data.Idea);
-    };
+    // set state when the data received
+    setData(data && data.Idea);
+  };
 
-    React.useEffect(() => {
+  React.useEffect(() => {
     getallproject();
   }, []);
-  console.log(data);
 
-  const deletIdea = async (id:string) => {
-    const data = await (
-      await fetch(
-        `http://localhost:3008/idea/${id}`,{
+  const deletIdea = async (id: string) => {
+    let result = confirm("هل انت متاكد؟؟!");
+    if (result == true) {
+      const data = await (
+        await fetch(`http://localhost:3008/idea/${id}`, {
           method: "delete",
           headers: {
             "Content-Type": "application/json",
-            "authorization": localStorage.getItem('token') as string,
+            authorization: localStorage.getItem("token") as string,
           },
-        }
-        )
-        ).json();
-        // we will come back to this to fix it!!!!!
-        // window.location.reload();
-        getallproject();
-        navigate("/Myideas")
+        })
+      ).json();
+    }
+    getallproject();
+    navigate("/Myideas");
   };
-  
+
   return (
     <div>
-       <nav>
+      <nav>
         <Nav />
       </nav>
       <Flex ml={40} mb={5} mt={20}>
-        <Box >
+        <Box>
           <Text fontSize="2xl" fontWeight="bold" textAlign={"right"} mr={40}>
             الافكار الخاصة بك
           </Text>{" "}
@@ -86,21 +81,21 @@ function Myideas() {
             onClick={() => navigate("/AddNewIdea")}
             m="10px"
             _hover={{ color: "white", backgroundColor: "#00ADBB" }}
-              color={"#00ADBB"}
-              bg={"none"}
-              border="1px solid #00ADBB"
+            color={"#00ADBB"}
+            bg={"none"}
+            border="1px solid #00ADBB"
           >
             إضافة فكرة جديد
           </Button>
         </Box>
       </Flex>
-{/* ------- this is where to code */}
+      {/* ------- this is where to code */}
 
-<SimpleGrid
+      <SimpleGrid
         borderColor={"blackAlpha.200"}
         minH={"60vh"}
         p={20}
-         borderRadius={"2xl"}
+        borderRadius={"2xl"}
         mx="auto"
         spacingX={10}
         spacingY={20}
@@ -109,64 +104,49 @@ function Myideas() {
         {/* {data== "" || "" ? "لا توجد لديك افكار " :""} */}
         {data.map((index) => (
           <div>
-          
-            <GridItem key={index.id} >
-              <Card 
-              h={320}
-              >
+            <GridItem key={index.id}>
+              <Card h={320}>
                 <CardHeader>
-                  <Text fontSize={30}  mb={4}>
-{index.title}                   
+                  <Text fontSize={30} mb={4}>
+                    {index.title}
                   </Text>
                   <hr />
-                  <Text>  </Text>
+                  <Text> </Text>
                 </CardHeader>
-                
+
                 <CardBody h={20}>
-                
                   <Text>{index.discription}</Text>
-                
                 </CardBody>
-                
+
                 <CardFooter>
-                   
-                <Box > 
-             <IconButton
-  color={'red'}
-  aria-label='delete'
-  icon={<DeleteIcon />}
-  onClick={()=>deletIdea(index.id)}
-/>
-  <Link to={`/ModifyIdea/${index.id}`}>
-<IconButton
-  mr={10}
-  color={'green'}
-  
-  
-  aria-label='edit'
-  icon={<EditIcon />}
-/>
-  </Link>
-</Box>
-
-                   
+                  <Box>
+                    <IconButton
+                      color={"red"}
+                      aria-label="delete"
+                      icon={<DeleteIcon />}
+                      onClick={() => deletIdea(index.id)}
+                    />
+                    <Link to={`/ModifyIdea/${index.id}`}>
+                      <IconButton
+                        mr={10}
+                        color={"green"}
+                        aria-label="edit"
+                        icon={<EditIcon />}
+                      />
+                    </Link>
+                  </Box>
                 </CardFooter>
-                
               </Card>
-              
             </GridItem>
-
-            
           </div>
-     ))}
+        ))}
       </SimpleGrid>
 
-{/* -------------------------------------------------------- */}
+      {/* -------------------------------------------------------- */}
 
-     
       <footer>
-        <Footer/>
-       </footer>
+        <Footer />
+      </footer>
     </div>
   );
 }
